requirements.txt
Flask==2.3.2
slackclient==2.9.4
slackeventsapi==3.0.1
langchain==0.0.263
openai==0.27.8